Require JPEG frame headers for image projection
